requirements.txt
# SignalForge Dependencies

# Core
python-dotenv>=1.0.1

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# API
fastapi>=0.109.0
# FIX: uvicorn[standard] should be in quotes to avoid shell errors in some terminals
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Data Processing
python-dateutil>=2.8.2
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0

# NLP & Text Processing
spacy>=3.7.2
textblob>=0.18.0

# Scheduling
apscheduler>=3.10.4

# Alerts
python-telegram-bot>=20.8

# Utilities
click>=8.1.7
colorama>=0.4.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.27.0

# Development
black>=24.2.0
flake8>=7.0.0
mypy>=1.8.0

# Dashboard
# Streamlit removed: the project now serves a static HTML UI via the FastAPI app at '/'
# Previously: streamlit>=1.30.0
